+++ src/datasources/data-frame/DataFrameDataSourceBase.ts
@@ -15,9 +15,9 @@ import {
     ValidDataFrameVariableQuery,
     DataFrameDataQuery,
     DataFrameVariableQuery,
-    Option,
-    ColumnFilter,
-    CombinedFilters
+    CombinedFilters,
+    ColumnOptions,
+    ColumnFilter
 } from './types';
 import { BackendSrv, TemplateSrv } from '@grafana/runtime';
 import { extractErrorInfo } from 'core/errors';
@@ -91,8 +91,8 @@ export abstract class DataFrameDataSourceBase<
         }
     }
 
-    public async getColumnOptionsWithVariables(filters: CombinedFilters): Promise<Option[]> {
-        return Promise.resolve([]);
+    public async getColumnOptionsWithVariables(filters: CombinedFilters): Promise<ColumnOptions> {
+        return Promise.resolve({ uniqueColumnsAcrossTables: [], commonColumnsAcrossTables: [] });
     }
 
     public async loadPartNumbers(): Promise<string[]> {